refactor: change file names and remove code repetition
